+++ ExilenceNextApp/src/components/simple-field/SimpleField.styles.ts
@@ -2,7 +2,7 @@ import { makeStyles } from '@material-ui/core';
 
 const useStyles = makeStyles(theme => ({
   root: {
-    marginBottom: theme.spacing(2)
+    marginBottom: theme.spacing(1)
   }
 }));