+++ ExilenceNextApp/src/store/domains/account.ts
@@ -55,6 +55,28 @@ export class Account implements IAccount {
   }
 
   @computed
+  get activeCharacter() {
+    const profile = this.activeProfile;
+    const accountLeague = this.accountLeagues.find(
+      l => l.leagueId === profile?.activeLeagueId
+    );
+    return accountLeague?.characters?.find(
+      ac => ac.name === profile?.activeCharacterName
+    );
+  }
+
+  get characters() {
+    const profile = this.activeProfile;
+    if (profile) {
+      return this.accountLeagues.find(
+        l => l.leagueId === profile.activeLeagueId
+      )?.characters;
+    } else {
+      return undefined;
+    }
+  }
+
+  @computed
   get activePriceLeague() {
     const profile = this.activeProfile;
     if (profile) {