+++ ExilenceNextApp/src/utils/dropdown.utils.ts
@@ -1,3 +1,4 @@
+// todo: make more generic
 export function getDropdownSelection<T extends { id: string }>(
   items: T[],
   activeId?: string
@@ -7,6 +8,19 @@ export function getDropdownSelection<T extends { id: string }>(
     return item.id;
   } else if (items.length > 0) {
     return items[0].id;
+  }
+  return '';
+}
+
+export function getNameDropdownSelection<T extends { name: string }>(
+  items: T[],
+  activeName?: string
+): string {
+  const item = activeName ? items.find(i => i.name === activeName) : undefined;
+  if (item) {
+    return item.name;
+  } else if (items.length > 0) {
+    return items[0].name;
   }
   return '';
 }