+++ ExilenceNextApp/src/utils/misc.utils.ts
@@ -1,3 +1,5 @@
 export function randomIntFromInterval(min: number, max: number) {
   return Math.floor(Math.random() * (max - min + 1) + min);
 }
+
+export const placeholderOption = 'None';